build(back): move Calculator to new folder
