+++ src/main/c/backend/domain-specific/calculator/Calculator.h
@@ -6,10 +6,10 @@
  * the layers of the backend and frontend using another group of
  * domain-specific models or DTOs (Data Transfer Objects).
  */
-#include "../../frontend/syntactic-analysis/AbstractSyntaxTree.h"
-#include "../../support/logging/Logger.h"
-#include "../../support/type/CompilerState.h"
-#include "../../support/type/ModuleDestructor.h"
+#include "../../../frontend/syntactic-analysis/AbstractSyntaxTree.h"
+#include "../../../support/logging/Logger.h"
+#include "../../../support/type/CompilerState.h"
+#include "../../../support/type/ModuleDestructor.h"
 #include <limits.h>
 #include <stdbool.h>
 
@@ -30,7 +30,7 @@ typedef ComputationResult (*BinaryOperator)(const int, const int);
 /** Arithmetic operations. */
 
 ComputationResult add(const int leftAddend, const int rightAddend);
-ComputationResult divide(const int dividend, const int divisor);
+ComputationResult divide(const int diidend, const int divisor);
 ComputationResult multiply(const int multiplicand, const int multiplier);
 ComputationResult subtract(const int minuend, const int subtract);
